Semi-drop Python 3.6 and 3.7
For the sake of DX, we are dropping the full support for Python 3.6 and
3.7, that is, the versions that have reached their end-of-life.

In order to fully support Django 3.2, the package can still be installed
on older Python versions, and the we also keep the syntax checks to
ensure the app runs on older Python versions. However, we are not
testing it in any way.